chore(federation): move all the rebasing code into its own module
