+++ apollo-federation/src/operation/tests/mod.rs
@@ -15,7 +15,7 @@ use crate::schema::position::ObjectTypeDefinitionPosition;
 use crate::schema::ValidFederationSchema;
 use crate::subgraph::Subgraph;
 
-fn parse_schema_and_operation(
+pub(super) fn parse_schema_and_operation(
     schema_and_operation: &str,
 ) -> (ValidFederationSchema, ExecutableDocument) {
     let (schema, executable_document) =
@@ -25,7 +25,7 @@ fn parse_schema_and_operation(
     (schema, executable_document)
 }
 
-fn parse_subgraph(name: &str, schema: &str) -> ValidFederationSchema {
+pub(super) fn parse_subgraph(name: &str, schema: &str) -> ValidFederationSchema {
     let parsed_schema =
         Subgraph::parse_and_expand(name, &format!("https://{name}"), schema).unwrap();
     ValidFederationSchema::new(parsed_schema.schema).unwrap()
@@ -1010,604 +1010,6 @@ scalar FieldSet
 }"#;
         let actual = normalized_operation.to_string();
         assert_eq!(expected, actual);
-    }
-}
-
-//
-// REBASE TESTS
-//
-#[cfg(test)]
-mod rebase_tests {
-    use apollo_compiler::name;
-    use indexmap::IndexSet;
-
-    use crate::operation::normalize_operation;
-    use crate::operation::tests::parse_schema_and_operation;
-    use crate::operation::tests::parse_subgraph;
-    use crate::operation::NamedFragments;
-    use crate::schema::position::InterfaceTypeDefinitionPosition;
-
-    #[test]
-    fn skips_unknown_fragment_fields() {
-        let operation_fragments = r#"
-query TestQuery {
-  t {
-    ...FragOnT
-  }
-}
-
-fragment FragOnT on T {
-  v0
-  v1
-  v2
-  u1 {
-    v3
-    v4
-    v5
-  }
-  u2 {
-    v4
-    v5
-  }
-}
-
-type Query {
-  t: T
-}
-
-type T {
-  v0: Int
-  v1: Int
-  v2: Int
-  u1: U
-  u2: U
-}
-
-type U {
-  v3: Int
-  v4: Int
-  v5: Int
-}
-"#;
-        let (schema, mut executable_document) = parse_schema_and_operation(operation_fragments);
-        assert!(
-            !executable_document.fragments.is_empty(),
-            "operation should have some fragments"
-        );
-
-        if let Some(operation) = executable_document.named_operations.get_mut("TestQuery") {
-            let normalized_operation = normalize_operation(
-                operation,
-                NamedFragments::new(&executable_document.fragments, &schema),
-                &schema,
-                &IndexSet::new(),
-            )
-            .unwrap();
-
-            let subgraph_schema = r#"type Query {
-  _: Int
-}
-
-type T {
-  v1: Int
-  u1: U
-}
-
-type U {
-  v3: Int
-  v5: Int
-}"#;
-            let subgraph = parse_subgraph("A", subgraph_schema);
-            let rebased_fragments = normalized_operation.named_fragments.rebase_on(&subgraph);
-            assert!(rebased_fragments.is_ok());
-            let rebased_fragments = rebased_fragments.unwrap();
-            assert!(!rebased_fragments.is_empty());
-            assert!(rebased_fragments.contains(&name!("FragOnT")));
-            let rebased_fragment = rebased_fragments.fragments.get("FragOnT").unwrap();
-
-            insta::assert_snapshot!(rebased_fragment, @r###"
-                    fragment FragOnT on T {
-                      v1
-                      u1 {
-                        v3
-                        v5
-                      }
-                    }
-                "###);
-        }
-    }
-
-    #[test]
-    fn skips_unknown_fragment_on_condition() {
-        let operation_fragments = r#"
-query TestQuery {
-  t {
-    ...FragOnT
-  }
-  u {
-    ...FragOnU
-  }
-}
-
-fragment FragOnT on T {
-  x
-  y
-}
-
-fragment FragOnU on U {
-  x
-  y
-}
-
-type Query {
-  t: T
-  u: U
-}
-
-type T {
-  x: Int
-  y: Int
-}
-
-type U {
-  x: Int
-  y: Int
-}
-"#;
-        let (schema, mut executable_document) = parse_schema_and_operation(operation_fragments);
-        assert!(
-            !executable_document.fragments.is_empty(),
-            "operation should have some fragments"
-        );
-        assert_eq!(2, executable_document.fragments.len());
-
-        if let Some(operation) = executable_document.named_operations.get_mut("TestQuery") {
-            let normalized_operation = normalize_operation(
-                operation,
-                NamedFragments::new(&executable_document.fragments, &schema),
-                &schema,
-                &IndexSet::new(),
-            )
-            .unwrap();
-
-            let subgraph_schema = r#"type Query {
-  t: T
-}
-
-type T {
-  x: Int
-  y: Int
-}"#;
-            let subgraph = parse_subgraph("A", subgraph_schema);
-            let rebased_fragments = normalized_operation.named_fragments.rebase_on(&subgraph);
-            assert!(rebased_fragments.is_ok());
-            let rebased_fragments = rebased_fragments.unwrap();
-            assert!(!rebased_fragments.is_empty());
-            assert!(rebased_fragments.contains(&name!("FragOnT")));
-            assert!(!rebased_fragments.contains(&name!("FragOnU")));
-            let rebased_fragment = rebased_fragments.fragments.get("FragOnT").unwrap();
-
-            let expected = r#"fragment FragOnT on T {
-  x
-  y
-}"#;
-            let actual = rebased_fragment.to_string();
-            assert_eq!(actual, expected);
-        }
-    }
-
-    #[test]
-    fn skips_unknown_type_within_fragment() {
-        let operation_fragments = r#"
-query TestQuery {
-  i {
-    ...FragOnI
-  }
-}
-
-fragment FragOnI on I {
-  id
-  otherId
-  ... on T1 {
-    x
-  }
-  ... on T2 {
-    y
-  }
-}
-
-type Query {
-  i: I
-}
-
-interface I {
-  id: ID!
-  otherId: ID!
-}
-
-type T1 implements I {
-  id: ID!
-  otherId: ID!
-  x: Int
-}
-
-type T2 implements I {
-  id: ID!
-  otherId: ID!
-  y: Int
-}
-"#;
-        let (schema, mut executable_document) = parse_schema_and_operation(operation_fragments);
-        assert!(
-            !executable_document.fragments.is_empty(),
-            "operation should have some fragments"
-        );
-
-        if let Some(operation) = executable_document.named_operations.get_mut("TestQuery") {
-            let normalized_operation = normalize_operation(
-                operation,
-                NamedFragments::new(&executable_document.fragments, &schema),
-                &schema,
-                &IndexSet::new(),
-            )
-            .unwrap();
-
-            let subgraph_schema = r#"type Query {
-  i: I
-}
-
-interface I {
-  id: ID!
-}
-
-type T2 implements I {
-  id: ID!
-  y: Int
-}
-"#;
-            let subgraph = parse_subgraph("A", subgraph_schema);
-            let rebased_fragments = normalized_operation.named_fragments.rebase_on(&subgraph);
-            assert!(rebased_fragments.is_ok());
-            let rebased_fragments = rebased_fragments.unwrap();
-            assert!(!rebased_fragments.is_empty());
-            assert!(rebased_fragments.contains(&name!("FragOnI")));
-            let rebased_fragment = rebased_fragments.fragments.get("FragOnI").unwrap();
-
-            let expected = r#"fragment FragOnI on I {
-  id
-  ... on T2 {
-    y
-  }
-}"#;
-            let actual = rebased_fragment.to_string();
-            assert_eq!(actual, expected);
-        }
-    }
-
-    #[test]
-    fn skips_typename_on_possible_interface_objects_within_fragment() {
-        let operation_fragments = r#"
-query TestQuery {
-  i {
-    ...FragOnI
-  }
-}
-
-fragment FragOnI on I {
-  __typename
-  id
-  x
-}
-
-type Query {
-  i: I
-}
-
-interface I {
-  id: ID!
-  x: String!
-}
-
-type T implements I {
-  id: ID!
-  x: String!
-}
-"#;
-
-        let (schema, mut executable_document) = parse_schema_and_operation(operation_fragments);
-        assert!(
-            !executable_document.fragments.is_empty(),
-            "operation should have some fragments"
-        );
-
-        if let Some(operation) = executable_document.named_operations.get_mut("TestQuery") {
-            let mut interface_objects: IndexSet<InterfaceTypeDefinitionPosition> = IndexSet::new();
-            interface_objects.insert(InterfaceTypeDefinitionPosition {
-                type_name: name!("I"),
-            });
-            let normalized_operation = normalize_operation(
-                operation,
-                NamedFragments::new(&executable_document.fragments, &schema),
-                &schema,
-                &interface_objects,
-            )
-            .unwrap();
-
-            let subgraph_schema = r#"extend schema @link(url: "https://specs.apollo.dev/link/v1.0") @link(url: "https://specs.apollo.dev/federation/v2.5", import: [{ name: "@interfaceObject" }, { name: "@key" }])
-
-directive @link(url: String, as: String, import: [link__Import]) repeatable on SCHEMA
-
-directive @key(fields: federation__FieldSet!, resolvable: Boolean = true) repeatable on OBJECT | INTERFACE
-
-directive @interfaceObject on OBJECT
-
-type Query {
-  i: I
-}
-
-type I @interfaceObject @key(fields: "id") {
-  id: ID!
-  x: String!
-}
-
-scalar link__Import
-
-scalar federation__FieldSet
-"#;
-            let subgraph = parse_subgraph("A", subgraph_schema);
-            let rebased_fragments = normalized_operation.named_fragments.rebase_on(&subgraph);
-            assert!(rebased_fragments.is_ok());
-            let rebased_fragments = rebased_fragments.unwrap();
-            assert!(!rebased_fragments.is_empty());
-            assert!(rebased_fragments.contains(&name!("FragOnI")));
-            let rebased_fragment = rebased_fragments.fragments.get("FragOnI").unwrap();
-
-            let expected = r#"fragment FragOnI on I {
-  id
-  x
-}"#;
-            let actual = rebased_fragment.to_string();
-            assert_eq!(actual, expected);
-        }
-    }
-
-    #[test]
-    fn skips_fragments_with_trivial_selections() {
-        let operation_fragments = r#"
-query TestQuery {
-  t {
-    ...F1
-    ...F2
-    ...F3
-  }
-}
-
-fragment F1 on T {
-  a
-  b
-}
-
-fragment F2 on T {
-  __typename
-  a
-  b
-}
-
-fragment F3 on T {
-  __typename
-  a
-  b
-  c
-  d
-}
-
-type Query {
-  t: T
-}
-
-type T {
-  a: Int
-  b: Int
-  c: Int
-  d: Int
-}
-"#;
-        let (schema, mut executable_document) = parse_schema_and_operation(operation_fragments);
-        assert!(
-            !executable_document.fragments.is_empty(),
-            "operation should have some fragments"
-        );
-
-        if let Some(operation) = executable_document.named_operations.get_mut("TestQuery") {
-            let normalized_operation = normalize_operation(
-                operation,
-                NamedFragments::new(&executable_document.fragments, &schema),
-                &schema,
-                &IndexSet::new(),
-            )
-            .unwrap();
-
-            let subgraph_schema = r#"type Query {
-  t: T
-}
-
-type T {
-  c: Int
-  d: Int
-}
-"#;
-            let subgraph = parse_subgraph("A", subgraph_schema);
-            let rebased_fragments = normalized_operation.named_fragments.rebase_on(&subgraph);
-            assert!(rebased_fragments.is_ok());
-            let rebased_fragments = rebased_fragments.unwrap();
-            // F1 reduces to nothing, and F2 reduces to just __typename so we shouldn't keep them.
-            assert_eq!(1, rebased_fragments.size());
-            assert!(rebased_fragments.contains(&name!("F3")));
-            let rebased_fragment = rebased_fragments.fragments.get("F3").unwrap();
-
-            let expected = r#"fragment F3 on T {
-  __typename
-  c
-  d
-}"#;
-            let actual = rebased_fragment.to_string();
-            assert_eq!(actual, expected);
-        }
-    }
-
-    #[test]
-    fn handles_skipped_fragments_within_fragments() {
-        let operation_fragments = r#"
-query TestQuery {
-  ...TheQuery
-}
-
-fragment TheQuery on Query {
-  t {
-    x
-    ... GetU
-  }
-}
-
-fragment GetU on T {
-  u {
-    y
-    z
-  }
-}
-
-type Query {
-  t: T
-}
-
-type T {
-  x: Int
-  u: U
-}
-
-type U {
-  y: Int
-  z: Int
-}
-"#;
-        let (schema, mut executable_document) = parse_schema_and_operation(operation_fragments);
-        assert!(
-            !executable_document.fragments.is_empty(),
-            "operation should have some fragments"
-        );
-
-        if let Some(operation) = executable_document.named_operations.get_mut("TestQuery") {
-            let normalized_operation = normalize_operation(
-                operation,
-                NamedFragments::new(&executable_document.fragments, &schema),
-                &schema,
-                &IndexSet::new(),
-            )
-            .unwrap();
-
-            let subgraph_schema = r#"type Query {
-  t: T
-}
-
-type T {
-  x: Int
-}"#;
-            let subgraph = parse_subgraph("A", subgraph_schema);
-            let rebased_fragments = normalized_operation.named_fragments.rebase_on(&subgraph);
-            assert!(rebased_fragments.is_ok());
-            let rebased_fragments = rebased_fragments.unwrap();
-            // F1 reduces to nothing, and F2 reduces to just __typename so we shouldn't keep them.
-            assert_eq!(1, rebased_fragments.size());
-            assert!(rebased_fragments.contains(&name!("TheQuery")));
-            let rebased_fragment = rebased_fragments.fragments.get("TheQuery").unwrap();
-
-            let expected = r#"fragment TheQuery on Query {
-  t {
-    x
-  }
-}"#;
-            let actual = rebased_fragment.to_string();
-            assert_eq!(actual, expected);
-        }
-    }
-
-    #[test]
-    fn handles_subtypes_within_subgraphs() {
-        let operation_fragments = r#"
-query TestQuery {
-  ...TQuery
-}
-
-fragment TQuery on Query {
-  t {
-    x
-    y
-    ... on T {
-      z
-    }
-  }
-}
-
-type Query {
-  t: I
-}
-
-interface I {
-  x: Int
-  y: Int
-}
-
-type T implements I {
-  x: Int
-  y: Int
-  z: Int
-}
-"#;
-        let (schema, mut executable_document) = parse_schema_and_operation(operation_fragments);
-        assert!(
-            !executable_document.fragments.is_empty(),
-            "operation should have some fragments"
-        );
-
-        if let Some(operation) = executable_document.named_operations.get_mut("TestQuery") {
-            let normalized_operation = normalize_operation(
-                operation,
-                NamedFragments::new(&executable_document.fragments, &schema),
-                &schema,
-                &IndexSet::new(),
-            )
-            .unwrap();
-
-            let subgraph_schema = r#"type Query {
-  t: T
-}
-
-type T {
-  x: Int
-  y: Int
-  z: Int
-}
-"#;
-
-            let subgraph = parse_subgraph("A", subgraph_schema);
-            let rebased_fragments = normalized_operation.named_fragments.rebase_on(&subgraph);
-            assert!(rebased_fragments.is_ok());
-            let rebased_fragments = rebased_fragments.unwrap();
-            // F1 reduces to nothing, and F2 reduces to just __typename so we shouldn't keep them.
-            assert_eq!(1, rebased_fragments.size());
-            assert!(rebased_fragments.contains(&name!("TQuery")));
-            let rebased_fragment = rebased_fragments.fragments.get("TQuery").unwrap();
-
-            let expected = r#"fragment TQuery on Query {
-  t {
-    x
-    y
-    z
-  }
-}"#;
-            let actual = rebased_fragment.to_string();
-            assert_eq!(actual, expected);
-        }
     }
 }
